requirements.txt
altgraph==0.17.4
keyboard==0.13.5
macholib==1.16.3
packaging==24.1
pillow==10.4.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyserial==3.5
setuptools==74.1.2
ttkthemes==3.2.2
